Allows QuickBooks Payments to work with OAuth v2 tokens
A fix to the _http function in Payments.php to allow QuickBooks Payments to work with the new OAuth v2 tokens.

Also, a quick and very dirty hack in IPP.php to allow Payments to renew the OAuth v2 token without relying on a QuickBooks Accounting call to do so.